HS-111 REFACT: check token 을 새로운 파일로
